RF: social tags checker
