+++ handlers/social_tags_test.go
@@ -7,82 +7,66 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"gopkg.in/h2non/gock.v1"
+	"github.com/xray-web/web-check-api/checks"
+	"github.com/xray-web/web-check-api/testutils"
 )
 
 func TestHandleGetSocialTags(t *testing.T) {
-	// t.Parallel()
-	tests := []struct {
-		name           string
-		urlParam       string
-		mockResponse   string
-		mockStatusCode int
-		expectedStatus int
-		expectedBody   map[string]interface{}
-	}{
-		{
-			name:           "Missing URL parameter",
-			urlParam:       "",
-			expectedStatus: http.StatusBadRequest,
-			expectedBody:   map[string]interface{}{"error": "missing URL parameter"},
-		},
-		{
-			name:           "Valid URL with social tags",
-			urlParam:       "http://example.com",
-			mockResponse:   `<html><head><title>Example Domain</title><meta name="description" content="Example description"><meta property="og:title" content="Example OG Title"></head><body></body></html>`,
-			mockStatusCode: http.StatusOK,
-			expectedStatus: http.StatusOK,
-			expectedBody: map[string]interface{}{
-				"title":              "Example Domain",
-				"description":        "Example description",
-				"keywords":           "",
-				"canonicalUrl":       "",
-				"ogTitle":            "Example OG Title",
-				"ogType":             "",
-				"ogImage":            "",
-				"ogUrl":              "",
-				"ogDescription":      "",
-				"ogSiteName":         "",
-				"twitterCard":        "",
-				"twitterSite":        "",
-				"twitterCreator":     "",
-				"twitterTitle":       "",
-				"twitterDescription": "",
-				"twitterImage":       "",
-				"themeColor":         "",
-				"robots":             "",
-				"googlebot":          "",
-				"generator":          "",
-				"viewport":           "",
-				"author":             "",
-				"publisher":          "",
-				"favicon":            "",
-			},
-		},
-	}
+	t.Parallel()
 
-	for _, tc := range tests {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			// t.Parallel()
-			defer gock.Off()
+	t.Run("Missing URL parameter", func(t *testing.T) {
+		t.Parallel()
+		req := httptest.NewRequest("GET", "/social-tag?url=", nil)
+		rec := httptest.NewRecorder()
 
-			if tc.urlParam != "" {
-				gock.New(tc.urlParam).
-					Reply(tc.mockStatusCode).
-					BodyString(tc.mockResponse)
-			}
+		HandleGetSocialTags(checks.NewSocialTags(nil)).ServeHTTP(rec, req)
 
-			req := httptest.NewRequest("GET", "/social-tags?url="+tc.urlParam, nil)
-			rec := httptest.NewRecorder()
-			HandleGetSocialTags().ServeHTTP(rec, req)
+		assert.Equal(t, http.StatusBadRequest, rec.Code)
+		var response KV
+		err := json.Unmarshal(rec.Body.Bytes(), &response)
+		assert.NoError(t, err)
+		assert.Equal(t, KV{"error": "missing URL parameter"}, response)
+	})
 
-			assert.Equal(t, tc.expectedStatus, rec.Code)
+	t.Run("Valid URL with social tags", func(t *testing.T) {
+		t.Parallel()
 
-			var responseBody map[string]interface{}
-			err := json.Unmarshal(rec.Body.Bytes(), &responseBody)
-			assert.NoError(t, err)
-			assert.Equal(t, tc.expectedBody, responseBody)
-		})
-	}
+		req := httptest.NewRequest("GET", "/social-tags?url=example.com", nil)
+		rec := httptest.NewRecorder()
+
+		HandleGetSocialTags(checks.NewSocialTags(testutils.MockClient(testutils.Response(http.StatusOK, []byte(`<html><head><title>Example Domain</title><meta name="description" content="Example description"><meta property="og:title" content="Example OG Title"></head><body></body></html>`))))).ServeHTTP(rec, req)
+
+		assert.Equal(t, http.StatusOK, rec.Code)
+
+		var responseBody KV
+		err := json.Unmarshal(rec.Body.Bytes(), &responseBody)
+		assert.NoError(t, err)
+		assert.Equal(t, KV{
+			"title":              "Example Domain",
+			"description":        "Example description",
+			"keywords":           "",
+			"canonicalUrl":       "",
+			"ogTitle":            "Example OG Title",
+			"ogType":             "",
+			"ogImage":            "",
+			"ogUrl":              "",
+			"ogDescription":      "",
+			"ogSiteName":         "",
+			"twitterCard":        "",
+			"twitterSite":        "",
+			"twitterCreator":     "",
+			"twitterTitle":       "",
+			"twitterDescription": "",
+			"twitterImage":       "",
+			"themeColor":         "",
+			"robots":             "",
+			"googlebot":          "",
+			"generator":          "",
+			"viewport":           "",
+			"author":             "",
+			"publisher":          "",
+			"favicon":            "",
+		}, responseBody)
+	})
+
 }
